bbu: battery backed block device cache driver (fast path)
Implement the i/o fast path (read, write, and barrier request handling)
of the bbu block device cache driver.  This borrows from the md-raid5
cache implementation.  The primary differnce is that bbu implements a
write-back cache in contrast to the write-through policy of md-raid5.  The
implication is that all read requests must check the cache for dirty
data before being submitted to the backing device.

Optimal operation is achieved when a read request does not hit dirty
data in the cache, or when the entire read request can be satisfied from
dirty data in the cache.  The third case (partial cache hit) is handled
by reading from the backing device and then merging in dirty data that
was present in the cache at the time of the initial request.

Signed-off-by: Dan Williams <dan.j.williams@intel.com>