e820: support for detecting and reserving ADR protected memory
ADR (Asynchronous DRAM refresh) is a mechanism to snapshot memory upon a
power loss event and place the DIMMs into self-refresh and typically
powered by a battery.  Unlike S3 an ADR event precludes the opportunity
to dump caches or save the system state.

ADR memory is advertised by e820 as major type AddressRangeReserved
(type-2) with the following bits set in the extended attributes: bit0
(Reserved), bit1 (AddressRangeNonVolatile), and bit4
(AddressRangeBatteryBacked).  Bit4 is pending ACPI committee approval
and indicates that memory is only protected for a duration on the order
of hours, i.e. not 100% non-volatile.

[minor: this patch also cleans up 'e820.map[i]' with a pointer]

Signed-off-by: Dan Williams <dan.j.williams@intel.com>