ntb: fixups for NTB to be able to used for bridging to the internet
We need to enable ARP. Also we need to allow checksum to be done for TCP
headers. Each NTB node must have unique MAC address as well. It does not
make sense to have hardcoded MAC addr for both ends of the NTB. One cannot
route packets with identical MAC.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>