bbu: battery backed block device cache driver (slow path cont'd)
Continue the implementation of the bbu driver slow path including the
block device to bbu_region registration interface and the initialization
code to start/restore the cache.  This cache uses memory protected by
the system ADR mechanism to store cache maintenance metadata (pfn to
sector associations / dirty state) and cache incoming write data.

The driver allows ADR memory to be partitioned into 64 caches.  Each of
these caches can be associated with a backing block device.  Currently
only md-raid5 devices are supported.  In general the device must be
already be using blk_queue_make_request() so that the bbu driver can
inject itself into the i/o path.  Associating a block device involves
communicating a uuid to the bbu cache.  Then when the block device is
started it registers with the bbu driver to restore its cache.  Assuming
a block device named md0 and a cache named cache0 the association can be
accomplished by:

	cat /sys/block/md0/md/uuid > /sys/class/bbu/cache0/uuid

Signed-off-by: Dan Williams <dan.j.williams@intel.com>