BBU: Fixing data corruption
The changes fix data corruption when the backing device is larger than the
BBU cache. Changes are added to prevent new data from being written to the
cache entries while old data are being flushed to the disks. Also, an
uninitialized flag variable was found for the blks in the cache entries.
This causes the BLK_F_UPTODATE flag not ever being removed. Once we go beyond
the BBU cache size can cover, new entires are being used with this flag
still set and thus BBU consider all cache up to date there on.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>