ioat: new channel reset scheme on Intel Atom S12x9 platforms
The Intel Atom S12x9 platforms changed the channel reset behavior.
It seems to do a reset similar to PCI FLR by resetting all the MSIX
registers. We have to re-init msix interrupts because of this. This
is only specific to this platform and is not expected to carry over to
the next generation.

Signed-off-by: Dave Jiang <dave.jiang@intel.com>