NTB: Xeon AER support
In Xeon NTB, AER appears as a vendor specific extended capability of PCI
Config space, and not the standard PCI AER capability.  Unfortunately,
this means it cannot use the Linux AER infrastructure and must do its
own detection and recovery.

When an AER event occurs, a link interrupt is encountered.  The driver
must now check on all link interrupts for AER events and Link changes.
For all correctable and non-fatal events, log the error and continue
running as normal.  If a fatal event is encountered, retrain the link
and attempt to recover.

Signed-off-by: Jon Mason <jon.mason@intel.com>