NTB: Heartbeat support
Add a heartbeat mechanism to determine if the remote system is
operational.  The local system periodically sends a doorbell to the
remote to let it know the system is functioning properly.  In the event
of a missed heartbeat, the driver will notify the transport layer that
the link is not functioning properly by sending a "link down"
notification.

Signed-off-by: Jon Mason <jon.mason@intel.com>