ioat: PM Support
Add power management support in the IOAT driver.  Adding this feature
resolves a known issue when attempting to suspend on systems with IOAT
enabled.  On those systems, IOAT would experience a DMA channel error
when attempting to resume due to the channels not being suspended
properly.  Currently, all channels errors cause panics, thus bringing
the system down.

Unfortunately, a hardware errata needs to be worked around on Xeon IOAT
devices.  The channel cannot be suspended or reset if there are active
XOR transactions.  To get around this, check to see if XOR is enabled on
the channel and the channel has pending transactions.  If so, then grab
the lock and wait for the queue to drain.  It might be more elegant to
see if there are any XOR operations pending in the list, but by that
time the queue will have drained anyway.

See BF509S in
http://www.intel.com/content/dam/www/public/us/en/documents/specification-updates/xeon-c5500-c3500-spec-update.pdf
See BT92 in
http://www.intel.com/content/dam/www/public/us/en/documents/specification-updates/xeon-e5-family-spec-update.pdf

Tested-by: Gary Hade <garyhade@us.ibm.com>
Signed-off-by: Jon Mason <jon.mason@intel.com>